# Improve replaceAll and replaceFirst from StringUtil.