- Rename `pattern` to `target` in order to avoid confusion with regex and to conform with standard `String.replace`.
- Avoid needless `StringBuilder` allocation.
- Avoid `String#substring`, which is no more `O(1)`.
- Add tests.