@codenameone 2019-11-15
Thanks!
The code doesn't compile in the CI. Notice that our internal code doesn't have access to all the features available to user code:

````bash
-do-compile:

    [mkdir] Created dir: /var/lib/jenkins/workspace/CodenameOne_PR-2952/cn1/CodenameOne/build/empty

    [mkdir] Created dir: /var/lib/jenkins/workspace/CodenameOne_PR-2952/cn1/CodenameOne/build/generated-sources/ap-source-output

    [javac] Compiling 602 source files to /var/lib/jenkins/workspace/CodenameOne_PR-2952/cn1/CodenameOne/build/classes

    [javac] /var/lib/jenkins/workspace/CodenameOne_PR-2952/cn1/CodenameOne/src/com/codename1/util/StringUtil.java:60: error: no suitable method found for append(String,int,int)

    [javac]         result.append(source, 0, index);

    [javac]               ^

    [javac]     method StringBuilder.append(char[],int,int) is not applicable

    [javac]       (argument mismatch; String cannot be converted to char[])

    [javac]     method StringBuilder.append(CharSequence,int,int) is not applicable

    [javac]       (argument mismatch; String cannot be converted to CharSequence)

    [javac] /var/lib/jenkins/workspace/CodenameOne_PR-2952/cn1/CodenameOne/src/com/codename1/util/StringUtil.java:66: error: no suitable method found for append(String,int,int)

    [javac]                 return result.append(source, start, source.length()).toString();

    [javac]                              ^

    [javac]     method StringBuilder.append(char[],int,int) is not applicable

    [javac]       (argument mismatch; String cannot be converted to char[])

    [javac]     method StringBuilder.append(CharSequence,int,int) is not applicable

    [javac]       (argument mismatch; String cannot be converted to CharSequence)

    [javac] /var/lib/jenkins/workspace/CodenameOne_PR-2952/cn1/CodenameOne/src/com/codename1/util/StringUtil.java:67: error: no suitable method found for append(String,int,int)

    [javac]             result.append(source, start, index);

    [javac]                   ^

    [javac]     method StringBuilder.append(char[],int,int) is not applicable

    [javac]       (argument mismatch; String cannot be converted to char[])

    [javac]     method StringBuilder.append(CharSequence,int,int) is not applicable

    [javac]       (argument mismatch; String cannot be converted to CharSequence)

    [javac] Note: Some input files use or override a deprecated API.

    [javac] Note: Recompile with -Xlint:deprecation for details.

    [javac] Note: Some input files use unchecked or unsafe operations.

    [javac] Note: Recompile with -Xlint:unchecked for details.

    [javac] Note: Some messages have been simplified; recompile with -Xdiags:verbose to get full output

    [javac] 3 errors

````