@Maaartinus 2019-11-15
@codenameone There's a single problem:
```
argument mismatch; String cannot be converted to CharSequence
```

This means that the main optimization can't be done in the most efficient way. The reason why I submitted the patch is the line
```
workingSource = workingSource.substring(idx + pattern.length());
```
which makes the complexity quadratic in terms of `source.length()`. I've seen such a thing taking minutes for a one megabyte source on a desktop PC (it'd be fine before Java 7u6).

> our internal code doesn't have access to all the features available to user code

Does it mean that I can use my version when I keep it in my own sources and it *will* work on the devices, right?

I could replace `result.append(source, start, end)` by `result.append(source.substring(start, end))` or by a loop, should I?