@codenameone 2019-11-16
Did you measure this? I'm not sure if it will perform well as some APIs are simulated.

I think the right solution for something like this might be something similar to the tokenize method approach where we implemented the core tokenize algorithm in native code. Alternatively this can also be implemented on top of the low level char arrays. I'm closing this PR for now, we can reopen it or create a new one with one of those approaches.