@Maaartinus 2019-11-16
> Did you measure this?

Now I did, but just on the desktop and without JMH.... but the results are more than clear:
[StringUtilBenchmark.java.txt](https://github.com/codenameone/CodenameOne/files/3854237/StringUtilBenchmark.java.txt)
```
length: 1.000 kilochars
new:     0.000005 seconds per replaceAll
old:     0.000036 seconds per replaceAll

length: 10.000 kilochars
new:     0.000054 seconds per replaceAll
old:     0.001848 seconds per replaceAll

length: 100.000 kilochars
new:     0.000205 seconds per replaceAll
old:     0.113226 seconds per replaceAll

length: 1000.000 kilochars
new:     0.002062 seconds per replaceAll
old:     8.998997 seconds per replaceAll
```
Sure, huge strings are rare on devices, but there's a factor of 30 even for 10k. And sure, the results on devices may differ, but not that much. And there are quite a few replacements in my string, but still....