@codenameone 2019-11-17
The simulator is meaningless in terms of performance as it's effectively JavaSE. Try this on the device. I'm betting it will have no impact there.